+++ examples/BarChart/CappedBars.tsx
@@ -10,7 +10,7 @@ const CappedBars = () => {
     {value: 30, label: 'Apr'},
   ];
   return (
-    <View>
+    <View style={{borderWidth:1}}>
       <BarChart
         data={data}
         barWidth={35}