+++ examples/BarChart/SimpleBlueBarsVerticalLines.tsx
@@ -13,7 +13,7 @@ const SimpleBlueBarsVerticalLines = () => {
     {value: 300, label: 'S'},
   ];
   return (
-    <View>
+    <View style={{borderWidth:1}}>
       <BarChart
         barWidth={22}
         noOfSections={3}