+++ examples/LineChart/SegmentedDataSet.tsx
@@ -5,25 +5,25 @@ import {DataSet} from '../../src/utils/types';
 
 const SegmentedDataSetChart = () => {
   const lineData = [
-    {value: 0, dataPointText: '0'},
-    {value: 10, dataPointText: '10'},
-    {value: 8, dataPointText: '8'},
-    {value: 58, dataPointText: '58'},
-    {value: 56, dataPointText: '56'},
-    {value: 78, dataPointText: '78'},
-    {value: 74, dataPointText: '74'},
-    {value: 98, dataPointText: '98'},
+    {value: 0},
+    {value: 10},
+    {value: 8},
+    {value: 58},
+    {value: 56},
+    {value: 78},
+    {value: 74},
+    {value: 98},
   ];
 
   const lineData2 = [
-    {value: 0, dataPointText: '0'},
-    {value: 20, dataPointText: '20'},
-    {value: 18, dataPointText: '18'},
-    {value: 40, dataPointText: '40'},
-    {value: 36, dataPointText: '36'},
-    {value: 60, dataPointText: '60'},
-    {value: 54, dataPointText: '54'},
-    {value: 85, dataPointText: '85'},
+    {value: 0},
+    {value: 20},
+    {value: 18},
+    {value: 40},
+    {value: 36},
+    {value: 60},
+    {value: 54},
+    {value: 85},
   ];
   const dataSet: Array<DataSet> = [
     {
@@ -44,10 +44,11 @@ const SegmentedDataSetChart = () => {
     },
   ];
   return (
-    <View>
+    <View style={{borderWidth:1}}>
       <LineChart
         dataSet={dataSet}
         height={250}
+        showValuesAsDataPointsText
         showVerticalLines
         spacing={44}
         initialSpacing={0}